Add building color status
I needed more than just the basic green for all successful builds or red for one or more failed builds.  I wanted the light to change to a different color when a build was active within the cctray view being watched.   Therefore I simply hacked the file to look first if something was building then change it blue otherwise the same behavior as before.   Simple change that I think adds a new element to this.   Perhaps this could be extended further to be configurable but I'm just learning Ruby and didn't want to jump in for anything more difficult than this.